Nest temporary class
